+++ src/platform_impl/macos/event.rs
@@ -1,4 +1,4 @@
-use std::os::raw::c_ushort;
+use std::{os::raw::c_ushort, path::PathBuf};
 
 use cocoa::{
     appkit::{NSEvent, NSEventModifierFlags},
@@ -18,6 +18,12 @@ use crate::{
 pub enum EventWrapper {
     StaticEvent(Event<'static, Never>),
     EventProxy(EventProxy),
+    MacSpecific(MacSpecificEvent),
+}
+
+#[derive(Debug)]
+pub enum MacSpecificEvent {
+    OpenFiles(Vec<PathBuf>),
 }
 
 #[derive(Debug, PartialEq)]