+++ src/platform_impl/macos/event_loop.rs
@@ -6,6 +6,7 @@ use std::{
     mem,
     os::raw::c_void,
     panic::{catch_unwind, resume_unwind, RefUnwindSafe, UnwindSafe},
+    path::PathBuf,
     process, ptr,
     rc::{Rc, Weak},
     sync::mpsc,
@@ -64,12 +65,18 @@ impl PanicInfo {
 pub struct EventLoopWindowTarget<T: 'static> {
     pub sender: mpsc::Sender<T>, // this is only here to be cloned elsewhere
     pub receiver: mpsc::Receiver<T>,
+
+    pub open_files_callback: RefCell<Option<Box<dyn FnMut(&[PathBuf])>>>,
 }
 
 impl<T> Default for EventLoopWindowTarget<T> {
     fn default() -> Self {
         let (sender, receiver) = mpsc::channel();
-        EventLoopWindowTarget { sender, receiver }
+        EventLoopWindowTarget {
+            sender,
+            receiver,
+            open_files_callback: RefCell::new(None),
+        }
     }
 }