+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 - On Android, support multi-touch.
 - On Wayland, extra mouse buttons are not dropped anymore.
 - **Breaking**: `MouseButton::Other` now uses `u16`.
+- On macOS, add `set_open_files_callback`.
 
 # 0.23.0 (2020-10-02)